+++ dandy/future/executor.py
@@ -1,18 +1,21 @@
 import concurrent.futures
 from time import sleep, time
 
-from typing_extensions import Callable
+from typing_extensions import Callable, TypeVar, Generic, Union
 
 from dandy.future.exceptions import FutureException
 
+
 ASYNC_EXECUTOR_RESULT_CHECK_INTERVAL = 0.001
+ASYNC_EXECUTOR = concurrent.futures.ThreadPoolExecutor()
+
+FutureResultType = TypeVar('FutureResultType')
 
 
 class AsyncExecutorFuture:
-    def __init__(self, func: Callable, *args, **kwargs):
-        self.executor = concurrent.futures.ThreadPoolExecutor()
-        self.future = self.executor.submit(func, *args, **kwargs)
-        self._result = None
+    def __init__(self, func: Callable[..., FutureResultType], *args, **kwargs):
+        self._future = ASYNC_EXECUTOR.submit(func, *args, **kwargs)
+        self._result: Union[FutureResultType, None] = None
         self._result_timeout = 0.0
         self._result_timer = 0.0
 
@@ -30,9 +33,9 @@ class AsyncExecutorFuture:
         return self._result_timeout > 0.0
 
     @property
-    def result(self):
+    def result(self) -> FutureResultType:
 
-        while not self.future.done():
+        while not self._future.done():
             sleep(ASYNC_EXECUTOR_RESULT_CHECK_INTERVAL)
 
             self._increment_result_timer()
@@ -41,14 +44,12 @@ class AsyncExecutorFuture:
                 raise FutureException(f'Future timed out after {self._result_timeout} seconds')
 
         try:
-            self._result = self.future.result()
+            self._result = self._future.result()
             return self._result
-        except concurrent.futures.TimeoutError as e:
-            raise FutureException(f'Future timed out: {e}')
-        except Exception as e:
-            raise FutureException(f'A problem occurred during async execution: {e}')
+        except concurrent.futures.TimeoutError:
+            raise FutureException(f'Future timed out')
         finally:
-            del self.future
+            self._future = None
 
     def set_timeout(self, seconds: int):
         self._result_timeout = seconds