+++ dandy/handler/handler.py
@@ -4,6 +4,7 @@ from abc import abstractmethod, ABCMeta
 from dandy.core.utils import json_default
 from dandy.debug.debug import DebugRecorder
 from dandy.debug.events import RunEvent, ResultEvent
+from dandy.future.executor import AsyncExecutorFuture
 
 
 class ProcessDebugABCMeta(ABCMeta):
@@ -63,3 +64,7 @@ class Handler(metaclass=ProcessDebugABCMeta):
     @abstractmethod
     def process(cls, **kwargs):
         ...
+
+    @classmethod
+    def process_to_future(cls, *args, **kwargs):
+        return AsyncExecutorFuture(cls.process, *args, **kwargs)